Expose http.Client in Pinning and Shaft struct
This commit adds a new field to the Pinning struct called Client, which is
a pointer to an http.Client instance. This field can be set using the new
Client() PinningOption function, allowing users to customize the http client
used by the pinning service. This change exposes more control over the underlying
http client, making it easier to configure and optimize network settings and
behavior for various use cases.